+++ src/monero.rs
@@ -10,7 +10,7 @@ use monero::util::key::{PrivateKey, PublicKey};
 use monero::Address;
 use monero::Amount;
 
-use std::fmt::{self, Debug, Display, Formatter};
+use std::fmt::{self, Debug};
 
 pub mod tasks;
 
@@ -74,16 +74,15 @@ impl std::str::FromStr for Monero {
 
     fn from_str(s: &str) -> Result<Self, Self::Err> {
         match s {
-            "Monero" => Ok(Monero),
+            "Monero" | "monero" | "xmr" => Ok(Monero),
             _ => Err(crate::consensus::Error::UnknownType),
         }
     }
 }
 
-impl Display for Monero {
-    fn fmt(&self, _f: &mut Formatter<'_>) -> fmt::Result {
-        println!("xmr");
-        Ok(())
+impl fmt::Display for Monero {
+    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+        write!(f, "Monero")
     }
 }